Further work on Descriptor struct: unit test, runtime init
Added default initialization for the default constructor and a constructor
from the integer type. Declaring all constructors constexpr.

Adding type cast operator to simplify the usage of Descriptor struct which
returns its integer member.

Adding a runtime init function in both DataDescription and generic
Descriptor to allow for use cases where the initializer string comes only
at runtime. The naming of the functions indicates that this is not a
compile-time feature and must be used with care.

Defining a type alias DataHeader::SubSpecificationType for the type of
member subSpecification of the DataHeader to allow generic use outside
the struct.