Bugfix: deleting additional data member in templated Descriptor struct
The Descriptor struct holds an n-byte descriptor as a union of an n-byte
char field and an appropriate integer type. The size of the Descriptor
must be equal to the number of bytes given as template argument.

This requirement was broken due to an additional data member for the
printer policy.

Fixes #373

Also adding appropriate unit tests and static assert to check data
structure sizes.