Enable a simulation unit test for TPC
Just until now, no simulation tests were being run.

Defining a cmake variable HAVESIMULATION that determines if we are compiling
in a sim enviroment (not DAQ) and use this to add one Geant4 and one Geant3
sim test for the TPC.

This is just a first step in extending our testing, but one which should detect
simple configuration errors etc. A more thourough/engineered solution will follow.