Second iteration on the DataProcessingLayer
* Add parallel / merge constructs to workflow

This is a first attempt at allowing expressing parallel processing and merging
of multiple streams of data in the declarative workflow.

It introduces two helper functions:

    o2::framework::parallel(...)

which allows duplicating and amending a template `DataProcessorSpec`, and

    o2::framework::mergeInputs(..)

which allows for duplicating InputSpec of a given device. Context information
for parallel computations (e.g. rank and number of parallel devices) can be
retrieved in the `ProcessCallback` via the `ParallelContext` service.

* Add time pipelining to the workflow

It's now possible to specify that a given `DataProcessor` will actually
be split in N devices, where each of the N devices takes one of the
timeslices produced by the previous step.

* Inputs now wrapped in InputRecord

Rather than exposing raw vector of DataRef to the algorithm writers, we
provide an InputRecord class which offers helper methods to access the
received inputs.

* ProcessingContext, InitContext, ErrorContext introduced

In order to simplify writing Algorithm callbacks, the AlgorithmSpec API has
been changed to take a "catch all" `*Context` object which exposes whatever is
relevant for the computation. This means that an algorithm spec will go from:

    AlgorithmSpec{
      [](InputRecord &, ServiceRegistry &, DataAllocator &){
      }
    }

to

    AlgorithmSpec{
      [](ProcessingContext &) {
      }
    }

* Improved internals

A lot of improvements to the internals have been done. In particular:

  - There is now one channel per device - device connection rather than one
    channel per exchanged datatype.
  - DataRelayer now handles more that 64 kind of inputs per device.
  - Increased number of tests and code coverage deriving from them.