Adding a frame collection parser and heartbeat sequence container
- adding a generic reverse parser for sequences of blocks composed
  of header - payload - trailer with variable payload length
- adding container class for multiple sequences of heartbeat frames
  with row and column access iterators

Further comments:
passing the access function to the iterator did not work because
the typedef for the access function is without the capture, so there
is no matching conversion.

Solution would be to use std::function but that's probably slow and
the function is called often. Can be checked later.