Fix non-compiling macros (#596)
This commit reviews all ROOT macros found in the O2 repository and fixes the ones
which were out of sync with the library code.
A few macros could not be fixed (since they called outdated functionality) and have hence been renamed as ".C_backup".

We are now in clean state and will henceforth check compilation/correct loading of macros during pull request checks. This will ensure that macros and library code stay consistent.