Only cluster of compatible ROFrame should be tracked together
Temporary solution for ITS tracking to overcome the problem of clusters from different ROFrames being
shipped in the single tree entry by the FairRunAna: multiple iterations over the clusters of single entry
will ensure that only clusters with the same ROFrameID (still may belong to different physical events) will
be tracked together.

Protection is added against reconstructing triggered data with cont. mode: if no cluster is selected
in non-empty event (because of the ROFrame mismatch) the Fatal will be produced