Consistent API, bugfix
Passing the TObject by reference to snapshot method to be consistent
with other specializations. The pointer const'ness has been removed
to be consistent with TMessageSerializer interface, keeping this
interface for the moment since a separate PR is open. Should be changed
later.

Bugfix for the single POD snapshot method, const'ness violating cast
removed

Adding test for TObject snapshot