[AOCSF-113] TrackReferences; Resolve issues in Stack
This commit introduces track reference handling in Simulation;
TrackReferences are managed by the Stack and can be added by Detectors
during processing.

The stack writes an indexed container of track references which can by queried
by trackID (corresponding to the kinematics track id in the MCTrack branch).

Introduction of a SimTrackStatus class encapsulating status information.

This commit also solves in inconsistency in the mother relationship of tracks
which was present.

Adding a unit test performing a more complex simulation and checking kinematics
and trackref properties.

Add a few track references to TPC, ITS for demonstration.