Bugfix: make pointer data types non-messageable
The type trait was classifying pointers as messageable because they are
trivially copyable. But this is not what we want. Pointers are easy to
pass to the DataAllocator API by mistake, catching this case with a
proper compile error now.

Adding a helper function to DataRefUtils to estract the payload size from
DataRef.