Correctly defining duration return type of getter
std::duration has been used the wrong way as template parameter and
only recent compilers complain about this. The return type is now
defined based on the rep and period member types which allows to
use either clock or duration types following the std::chrono concept.

We have discussed a conceptual question: both get<Clock> and get<Duration>
return a duration, and in the case of a clock type it's the subtype while
in the latter it's the type itself. For the moment it's convenient that
way, however subject to change if it turns out to be confusing.