Unifying the return type of the DPL input API
This reflects the outcome of the discussion following PR #973. The result for
get<T> (T not being a pointer) is either
- a smart pointer for type T, or
- an object providing iterator and begin/end methods to access elements

Special cases like get<const char*> and get<std::string> are preserved.

The smart-pointer case is fully covered by unique_ptr with the custom deleter
already available. We can later consider to factor out to separate file.

As for containers, currently only serialized std containers of types with
ROOT dictionary are supported. A general container_ref type implementation is
in progress.