Introducing background/signal merging
This commit provides a major new feature for the hit processing
stage. We are now able to overlay hits from different chains/files (in digitization).

This is mainly achieved through

* Extension of the HitProcessingManager/RunContext, with the new ability
  to take background/signal filenames and to sample composite collisions starting
  from elementary events

The new feature is demonstrated in the digitizer workflow example for the TPC

* digitizer-workflow -b --simFile background.root [--simFileS signal.root]

In order to make this work, some functions/structures in the TPC needed to be generalized:

* passing down labels is now done using directly the MCCompLabel class in order
  to correctly propagate the sourceID (previously not done)

* more generic TPCHitGroupID

* adapted version of the getHits() function to filter hits for any drifttime

This commit is part of [AOCSF-111]