Introducing collaborative simulation using FairMQ devices
This is a commit providing (the first version of) a major development
It introduces parallel simulation capabilities, in which
multiple simulation workers collaborate on the simulation of
a set of events. The system is implemented using FairMQ devices
communicating via messages.

The key components of the system are

1) O2PrimaryServerDevice
------------------------

A server device, generating events (primaries) using some
event generator and serving chunks of primaries to
simulation workers.

2) O2SimDevice:
---------------

A worker device actually performing a VMC simulation on a chunk
of primaries. The chunk can be smaller than a full event so that
multiple workers can collaboratively simulate a big event.
Simulation products (hits) are forwarded to a merger device which
is responsible to do the IO

3) O2HitMergerDevice:
---------------------

The device aggregating hits from the workers and doing IO to
a ROOT file.

Other notes:

- A driver application `o2sim_parallel` is provided that
  sets up the system behind the scenes and which feels like
  the standard o2sim executable but parallelized:

    o2sim_parallel -m PIPE ITS TPC -g pythia8 -e TGeant4 -n 20

- A couple of important changes accompagny this commit

  *) introduction of an O2MCApplication customizing certain aspects
     of FairMCApplication

  *) addition of O2RunSim which sets up O2MCApplication

  *) addition of some members to the base Detector class, amending
     it to prepare hit messages for a given FairMQChannel

  *) modification of the o2sim.C macro in order to be able
     to compile for the old o2sim executable as well as the new
     parallel system

- The fact that we process sub-events (aka primary chunks) might
  make the handling of MCTruth etc more complex in the future