Increasing timeout for DPL unit test workflows from 30 to 60 sec
Looking at one of the latest log files of a failing DPL workflow unit test
on the macos CI it seems that the failing test has been aborted in the middle
of the processing. Maybe the startup for some reason takes too long time and
then the test can not properly finish within the timeout.