Code reduction: Automatically implement CloneModule for detectors
This commit provides a more "framework" like approach to implementing
CloneModule for detectors. From now on this function can no longer be forgotten
and the requirements for G4-MT are automatically satisfied. This works
by using the templated DetImpl class.

This commit follows the "dont-repeat-yourself" principle. I plan to extend
it to other common functions in simulation detector code.

Some other minor cleanup is done (removal of TClonesArray forward decl)