DPL: support dangling inputs
Those are inputs which do not have a counterpart in the workflow
spec definition but which can be treated automatically by either
the framework (e.g. in the case of conditions) or by the deployment
system (e.g. in the case of deploying on the O2 farm, we should get
the input directly from the TimeframeBuilder.