DPL simulation workflow : Tutorial Part 2
This is part of a step-by-step tutorial of how to add a new digitizer
to the DPL digitizer workflow. It uses the example of TOF to build the solution
in a series of commits/parts.

This is part 2 which interfaces the actual digitization and transforms hits
to digits.

It adds the following
- add an instance of the o2::tof::Digitizer class which we call
  from the DPL processor function
  (any detector will typically have such a class which implements
   the actual transformation of hits to digits)

- instances of digit and label containers

- read the hits from the input chains

- loop over all collisions and collision parts and pass
  in each case the hits to the digitizer

Note that the logic in the DPL processor function is similar
to the one in typical DigitizerTasks. The main difference is
that we no longer use the FairRootManager to retrieve collisions times
and input.

Note that digits are not yet stored to file. This will be part of
the next step: Forwarding hits to another component (such as one doing IO).

The code here should be seen to describe the rough logic and how
to reuse existing algorithmic components. It is expected that some stuff
will be externalized into common service components (such as retrieval of hits, etc).