DPL simulation workflow : Tutorial Part 3
This is part of a step-by-step tutorial of how to add a new digitizer
to the DPL digitizer workflow. It uses the example of TOF to build the solution
in a series of commits/parts.

This is part 3 which provides a consumer processor for TOF digits in terms
of a simple FileSink (TOFDigitWriter).

It adds the following
- add the TOFDigitWriter processor, define its input, and add it to
  the DPL workflow (different consumers such as a clusterizer should look similar)

- writing digits to a configurable file (e.g. tofdigits.root)

- add an output to the TOFDigitizer where digits will be sent

- accumulate all digits in the TOFDigitizer and send/snapshot them

With this part we already have a decently working solution that transforms
hits into digits and stores them into a file, all by using DPL components.

As previously, the code offers many opportunities to factor out stuff into common services.