Bugfix: order of clusters written to CRU raw pages is now fixed, sorting MC labels
While raw pages have been allocated according to the correct numbers of clusters
per CRU, clusters have been written in the wrong sequence, simply from the first
to the last cluster in the input, rather then the clusters actually belonging to
a CRU.

This bug resultet in memory corruption in the HardwareClusterDecoder so we keep
this commit separate to reproduce the failure and improve robustness of the decoder
later.

Also implementing the sorting of MC labels in the cluster converter spec, but its
not yet activated.