DPL: prevent InputSpec default constructor
As we move closer to have InputSpec as a veritable "matcher" object
rather than a set of variables to be equal compared, I need to tight
the access to the member variables. The first adiabatic step in such a
direction is to make sure an InputSpec is built all at once, and not
incrementally, as that ability will be lost once we move to use
a DataDescriptorMatcher to describe the InputSpec internal State.