DPL: make relaying backend query based
This refactors TimesliceIndex and DataRelayer to be based
on `DataDescriptorMatcher` for the caching logic. This
really means we now do support queries on the data, both
in terms of description and in terms of time ranges, if
we had a way to specify such queries from the InputSpec API.

Hip Hip Hurrah!