Make sure geometry file is available in standard way
Little fix for parallel simulation, where each worker
creates a geometry file + a PID identifier.
We now make sure that the standard file (or a link) O2geometry.root
is created, which is expected by the processing chain further up.

With this, we can now run o2sim_parallel + digitizer-workflow.
Thanks to @davidrohr.