Implemented customization of help messages of the DPL driver app
An optional argument can be specified to the `--help,-h` option, to create
different versions of the help printout:
- `full`: the full printout with executor, workflow and processor options
- `short`: executor and workflow options, as well as a summary of the processors
- `workflow`: workflow options only
- `executor`: executor options only
- by specifying a valid name for a process which defines option, its options are printed

Default behaviour is `short`.

The definition of processors might depend on the specified workflow options. The help
message is adjusted to workflow options if specified in addition to the `--help` option.