DPL: reduce verbosity to define Algorithms
This introduces helpers which will hopefully greatly reduces the typing
required to implement Algorithms, when using the lambda based API of
AlgorithmSpec.

As you can read in the documentation with this the user simply needs
to pass a lambda where he specifies as inputs any of InputRecord,
DataAllocator, ConfigParamService, or any of the services hanging
from the ServiceRegistry, e.g. Monitoring. The actual extraction
of those classes happens via the template magic, leaving the user with
just:

  AlgorithmSpec{
    adaptStateless([](InputRecord &inputs, DataAllocator& output){
    ...
  })}

a similar adaptor is available for the stateful versions (i.e. one which
would have taken an InputRegistry).

In the future we might embed this directly into the AlgorithmSpec
constructor, without the need of an helper.