DPL Driver: speedup metric handling
Old code was recalculating the metrics to display for every iteration,
rather than doing it only when a new metric was added like now. While
this worked well for a few metrics, as we increase the number of them,
this becomes an obvious performance bottleneck.

While in production (i.e. when running on O2 Control / DDS) this part of
the code should not be exercised, several people use the DPL driver code
as convenience way to develop / debug issues and this was affecting
performance there in a significant way.