o2sim: stability fix for 0 events
There was a problem when the user asked for 0 events
to be simulated, since in this case the data merger device never
received any data and would hang.
This problem is now fixed by letting the merger know how many
events to expect (from analysing the configuration). This probably makes sense
but leads to a tighter coupling. To be reviewed.

Note that the 0 event case is useful since we want to expect
the geometry created during initialization etc.