Generator-level implementation of embedding into background (#1674)
* Generator-level implementation of embedding into background

This commit implements a basic mechanism to allow embedding
of signal events on pre-generated background events.

The currently-supported background source is file based,
namely the user needs to specify the filename (o2sim.root)
that contains the background events. The primary generator
of the signal events will force the primary interaction
vertex to be the same as the one of the background event.
Background events are sampled one-by-one until all events
have been used. At that point the events start to be reused.

User can select the embedding mode via o2sim command-line:
o2sim --embedIntoFile o2sim.bkg.root

* o2 implementation of FairMCEventHeader

Base implementation of the O2 version of the FairMCEventHeader.
Currently it stores information about embedding generation, namely
the name of the file used for the embedding and the event index.

* Update DetectorSimulation.md