Extend the external generator (extgen) concept to any FairGenerator
Modifies the behaviour of o2sim with the external generator option -g extgen.
It now expects that the external configuration macro defined via --extGenFile returns a FairGenerator* instead of a TGenerator* as it was before.
This allows for a more generic handling of customised external generators.

The TGenerator interface functionality is preserved, as everything can be defined in the configuration macro. Please see for example

share/external/tgenerator.C
share/external/pythia6.C

This change simplifies significantly the code and provides the user with more flexibility.