Introducing context objects to MessageContext
The MessageContext now supports polymorphic context objects derived from a common
interface. The current use case is to create a context object holding also the
intance of a span which is returned by reference by the DataAllocator API.

The MessageContext has methods to create messages vis the device proxy, the
actual creation of payload messages is controlled by the context object.

The concept can be extended in order to drop the different message contexts, e.g.
StringContext, RootObjectContext, in favor of the context objects handled by a
single message context.